# docs/swagger.yaml
basePath: /
definitions:
  genprotos.GetAllUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/genprotos.GetByIdUserResponse'
        type: array
    type: object
  genprotos.GetByIdUserResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      profile_picture:
        type: string
      role:
        type: string
    type: object
  genprotos.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  genprotos.RegisterUserRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      profile_picture:
        type: string
    type: object
  genprotos.UpdateUserForUser:
    properties:
      bio:
        type: string
      full_name:
        type: string
      profile_picture:
        type: string
    type: object
  genprotos.UpdateUserRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      profile_picture:
        type: string
    type: object
  genprotos.UpdateUserResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      profile_picture:
        type: string
    type: object
  handler.changePass:
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
    type: object
  handler.resetPass:
    properties:
      newPassword:
        type: string
      resetToken:
        type: string
    type: object
info:
  contact: {}
  description: auth service API
  title: auth service API
  version: "1.0"
paths:
  /admin/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successfully
          schema:
            type: string
        "400":
          description: Error while deleting user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get By ID Successful
          schema:
            $ref: '#/definitions/genprotos.GetByIdUserResponse'
        "400":
          description: Error while retrieving user
          schema:
            type: string
        "404":
          description: User Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get User By ID
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "400":
          description: Error while updating user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - Admin
  /admin/all:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users with pagination
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of Users
          schema:
            $ref: '#/definitions/genprotos.GetAllUsersResponse'
        "400":
          description: Error while retrieving users
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - Admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login Successfully
          schema:
            type: string
        "400":
          description: Error while logging in
          schema:
            type: string
        "404":
          description: User Not Found
          schema:
            type: string
      summary: Login User
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Create Successfully
          schema:
            type: string
        "400":
          description: Error while creating user
          schema:
            type: string
      summary: Register User
      tags:
      - Auth
  /user:
    delete:
      consumes:
      - application/json
      description: Delete an existing Profil
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "400":
          description: Error while deleting user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Profile
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get a user Profil
      produces:
      - application/json
      responses:
        "200":
          description: Get Profil Successful
          schema:
            $ref: '#/definitions/genprotos.GetByIdUserResponse'
        "400":
          description: Error while retrieving user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.UpdateUserForUser'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            $ref: '#/definitions/genprotos.UpdateUserResponse'
        "400":
          description: Error while updating user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - User
  /user/change-password:
    post:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Change Password
        in: body
        name: ChangePass
        required: true
        schema:
          $ref: '#/definitions/handler.changePass'
      produces:
      - application/json
      responses:
        "200":
          description: Password Changed Successfully
          schema:
            type: body
        "400":
          description: Error while changing password
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - User
  /user/forgot-password:
    post:
      consumes:
      - application/json
      description: Initiate forgot password process
      produces:
      - application/json
      responses:
        "200":
          description: Forgot Password Initiated Successfully
          schema:
            type: body
        "400":
          description: Error while initiating forgot password
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Forgot Password
      tags:
      - User
  /user/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password
      parameters:
      - description: Reset Password
        in: body
        name: ResetPass
        required: true
        schema:
          $ref: '#/definitions/handler.resetPass'
      produces:
      - application/json
      responses:
        "200":
          description: Password Reset Successfully
          schema:
            type: string
        "400":
          description: Error while resetting password
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Reset Password
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"